* Many conflicts with vehiclepatch
* updated VehicleGUID from uint64 to ObjectGUID

+++ src/shared/Threading.h
@@ -26,7 +26,6 @@
 
 namespace ACE_Based
 {
-
     class Runnable
     {
         public:
@@ -45,7 +44,7 @@ namespace ACE_Based
                 }
             }
         private:
-            ACE_Atomic_Op<ACE_Thread_Mutex, int> m_refs;
+            ACE_Atomic_Op<ACE_Thread_Mutex, long> m_refs;
 
             bool m_deleted;
     };
@@ -110,6 +109,5 @@ namespace ACE_Based
             //use this object to determine current OS thread priority values mapped to enum Priority{}
             static ThreadPriority m_TpEnum;
     };
-
 }
 #endif
